+++ cmd/swiper/main.go
@@ -8,10 +8,10 @@ import (
 	"runtime"
 	"runtime/pprof"
 
-	"swiper/internal/batch"
-	"swiper/internal/config"
-	"swiper/internal/extractor"
-	"swiper/internal/scanner"
+	"github.com/onedusk/swiper/internal/batch"
+	"github.com/onedusk/swiper/internal/config"
+	"github.com/onedusk/swiper/internal/extractor"
+	"github.com/onedusk/swiper/internal/scanner"
 )
 
 func main() {